@@ -19,7 +19,7 @@ version: 1.0.0+1
 
 environment:
   #sdk: '>=2.12.0 <3.0.0'
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
