@@ -2,14 +2,14 @@ import 'package:flutter/material.dart';
 
 class Responsive extends StatelessWidget {
   final Widget mobile;
-  final Widget tablet;
+  final Widget? tablet;
   final Widget desktop;
 
   const Responsive({
-    Key key,
-    @required this.mobile,
+    Key? key,
+    required this.mobile,
     this.tablet,
-    @required this.desktop,
+    required this.desktop,
   }) : super(key: key);
 
 // This size work fine on my design, maybe you need some customization depends on your design
@@ -34,7 +34,7 @@ class Responsive extends StatelessWidget {
     }
     // If width it less then 1100 and more then 850 we consider it as tablet
     else if (_size.width >= 850 && tablet != null) {
-      return tablet;
+      return tablet!;
     }
     // Or less then that we called it mobile
     else {
