@@ -6,8 +6,8 @@ import '../../../constants.dart';
 
 class FileInfoCard extends StatelessWidget {
   const FileInfoCard({
-    Key key,
-    @required this.info,
+    Key? key,
+    required this.info,
   }) : super(key: key);
 
   final CloudStorageInfo info;
@@ -27,11 +27,11 @@ class FileInfoCard extends StatelessWidget {
                 height: 40,
                 width: 40,
                 decoration: BoxDecoration(
-                  color: info.color.withOpacity(0.1),
+                  color: info.color!.withOpacity(0.1),
                   borderRadius: const BorderRadius.all(Radius.circular(10)),
                 ),
                 child: SvgPicture.asset(
-                  info.svgSrc,
+                  info.svgSrc!,
                   color: info.color,
                 ),
               ),
@@ -39,7 +39,7 @@ class FileInfoCard extends StatelessWidget {
             ],
           ),
           Text(
-            info.title,
+            info.title!,
             maxLines: 1,
             overflow: TextOverflow.ellipsis,
           ),
@@ -55,15 +55,15 @@ class FileInfoCard extends StatelessWidget {
                 style: Theme
                     .of(context)
                     .textTheme
-                    .caption
+                    .caption!
                     .copyWith(color: Colors.black54),
               ),
               Text(
-                info.totalStorage,
+                info.totalStorage!,
                 style: Theme
                     .of(context)
                     .textTheme
-                    .caption
+                    .caption!
                     .copyWith(color: Colors.black54),
               ),
             ],
@@ -76,13 +76,13 @@ class FileInfoCard extends StatelessWidget {
 
 class ProgressLine extends StatelessWidget {
   const ProgressLine({
-    Key key,
+    Key? key,
     this.color = primaryColor,
-    @required this.percentage,
+    required this.percentage,
   }) : super(key: key);
 
-  final Color color;
-  final int percentage;
+  final Color? color;
+  final int? percentage;
 
   @override
   Widget build(BuildContext context) {
@@ -92,13 +92,13 @@ class ProgressLine extends StatelessWidget {
           width: double.infinity,
           height: 5,
           decoration: BoxDecoration(
-            color: color.withOpacity(0.1),
+            color: color!.withOpacity(0.1),
             borderRadius: BorderRadius.all(Radius.circular(10)),
           ),
         ),
         LayoutBuilder(
           builder: (context, constraints) => Container(
-            width: constraints.maxWidth * (percentage / 100),
+            width: constraints.maxWidth * (percentage! / 100),
             height: 5,
             decoration: BoxDecoration(
               color: color,
