@@ -8,7 +8,7 @@ import 'file_info_card.dart';
 
 class MyFiles extends StatelessWidget {
   const MyFiles({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
@@ -48,7 +48,7 @@ class MyFiles extends StatelessWidget {
 
 class FileInfoCardGridView extends StatelessWidget {
   const FileInfoCardGridView({
-    Key key,
+    Key? key,
     this.crossAxisCount = 4,
     this.childAspectRatio = 1,
   }) : super(key: key);
