@@ -1,8 +1,8 @@
 import 'package:admin/model/ticket_static.dart';
 
 class ScreenArguments {
-  final String ticketId;
-  final Ticket ticket;
+  final String? ticketId;
+  final Ticket? ticket;
 
   ScreenArguments({this.ticketId, this.ticket});
 }
