@@ -2,9 +2,9 @@ import 'package:admin/constants.dart';
 import 'package:flutter/material.dart';
 
 class OwnTextField extends StatelessWidget{
-  final String hintText;
-  final String labelText;
-  final TextEditingController controller;
+  final String? hintText;
+  final String? labelText;
+  final TextEditingController? controller;
   OwnTextField({this.hintText, this.labelText, this.controller});
 
 
@@ -39,11 +39,11 @@ class OwnTextField extends StatelessWidget{
 }
 
 class OwnTextFieldWithIcons extends StatelessWidget{
-  final String hintText;
-  final String labelText;
-  final Widget suffixIcon;
-  final IconData prefixIcon;
-  final TextEditingController controller;
+  final String? hintText;
+  final String? labelText;
+  final Widget? suffixIcon;
+  final IconData? prefixIcon;
+  final TextEditingController? controller;
   OwnTextFieldWithIcons({this.hintText, this.labelText, this.suffixIcon, this.controller, this.prefixIcon});
 
   @override
@@ -82,9 +82,9 @@ class OwnTextFieldWithIcons extends StatelessWidget{
 }
 
 class OwnBigTextField extends StatelessWidget{
-  final String hintText;
-  final String labelText;
-  final TextEditingController controller;
+  final String? hintText;
+  final String? labelText;
+  final TextEditingController? controller;
   OwnBigTextField({this.hintText, this.labelText, this.controller});
 
 
