@@ -1,5 +1,5 @@
 class RecentFile {
-  final String icon, title, date, size;
+  final String? icon, title, date, size;
 
   RecentFile({this.icon, this.title, this.date, this.size});
 }
